+++ src/Gnugat/Redaktilo/Command/LineInsertAboveCommand.php
@@ -14,7 +14,7 @@ namespace Gnugat\Redaktilo\Command;
 use Gnugat\Redaktilo\Converter\LineContentConverter;
 
 /**
- * Inserts the given addition in the given file at the given location.
+ * Inserts the given addition in the given file above the given location.
  */
 class LineInsertAboveCommand implements Command
 {